major overhaul of ai internals
This is the first significant overhaul of the ai engine since 1994.

At that time, I serendipitously stumbled upon the concept of piece tables,
which placed a value on each piece and king by virtue of its location on the
board. It was really simple, and pretty smart, but too limited to know
about any kind of formations or patterns of play, depending too heavily
on recursive fortune-telling.

So I finally introduced the concept of "formations." Now it's possible to
specify not just an individual piece, but a pattern specifier, which will
trigger a (location-dependent!) score when it matches. It's tight, hand-
optimized, and freaky fast.

So now I can finally experiment with more intelligent patterns, and even
with tweaking scores and pitting different weights and inputs against each
other, like I've wanted to do for years.